Fix warning
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@15458 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/cmp.cpp
@@ -4490,7 +4490,7 @@ static jrd_nod* pass1_update(thread_db* tdbb,
 	if (trigger)
 	{
 		bool user_triggers = false;
-		for (int i = 0; i < trigger->getCount(); i++)
+		for (size_t i = 0; i < trigger->getCount(); i++)
 		{
 			if (!(*trigger)[i].sys_trigger)
 			{
